Continued with the upscaling studies

See merge request hpcsoftware/Peano!70

+++ python/exahype2/postprocessing/plot-scaling.py
@@ -13,6 +13,10 @@ if __name__ == "__main__":
   parser = argparse.ArgumentParser(description='ExaHyPE 2 scaling plotter')
   parser.add_argument("file",   help="filename of the file to parse (should be a tar.gz file)")
   parser.add_argument("-v", "--verbose", help="increase output verbosity", default=False)
+  parser.add_argument("-gc", "--grid-construction", dest="grid_construction", help="plot grid construction time, too", type=bool, default=False)
+  parser.add_argument("-li", "--last-iteration", dest="last_iteration", help="measure only the last iteration", type=bool, default=False)
+  parser.add_argument("-p", "--pattern", dest="file_pattern", help="define pattern that has to be in the filename", default="")
+  parser.add_argument("-log", dest="log", help="plot with logarithmic axes", default=False)
   args = parser.parse_args()
 
   open_mode = ""  
@@ -28,29 +32,33 @@ if __name__ == "__main__":
   
   data_points = []
   for data_file in data_files:
-    print( "parse " + data_file )
-    tar.extract( data_file )
-    new_data = exahype2.postprocessing.PerformanceDataPoint(data_file,args.verbose)
-    if new_data.valid:
-      data_points.append( new_data ) 
-    os.remove( data_file )
+    if args.file_pattern in data_file:
+      if args.verbose:
+        print( "parse " + data_file )
+      tar.extract( data_file )
+      new_data = exahype2.postprocessing.PerformanceDataPoint(data_file,args.verbose)
+      if new_data.valid:
+        data_points.append( new_data ) 
+      os.remove( data_file )
+    elif args.verbose:
+      print( "ignore " + data_file + " as file name does not match pattern")
 
   plt.clf()
 
-  (x_data, y_data) = exahype2.postprocessing.extract_grid_construction_times( data_points )    
-  plt.plot( x_data, y_data, label="total grid construction (incl. initial lb)" )
+  if args.grid_construction:
+    (x_data, y_data) = exahype2.postprocessing.extract_grid_construction_times( data_points )    
+    plt.plot( x_data, y_data, label="total grid construction (incl. initial lb)" )
 
-  (x_data, y_data) = exahype2.postprocessing.extract_times_per_step( data_points )    
+  (x_data, y_data) = exahype2.postprocessing.extract_times_per_step( data_points, args.last_iteration )    
   plt.plot( x_data, y_data, label="time per time step" )
 
   plt.plot( x_data, exahype2.postprocessing.linear_runtime_trend_line(x_data,y_data), "--", label="linear trend" )
   
   plt.xlabel( "Threads" )
   plt.ylabel( "Time [t]=s" )
-  plt.xscale( "log", basex=2 )
-  #plt.yscale( "log", basey=2 )
-  #plt.xscale( "log" )
-  plt.yscale( "log" )
+  if args.log:
+    plt.xscale( "log", basex=2 )
+    plt.yscale( "log" )
   xtics   = [ 1 ]
   xlabels = [ "serial" ]
   while xtics[-1] < x_data[-1]:
